+++ java/src/itp/comptypo/PhraseKerningCalibration.java
@@ -39,7 +39,7 @@ public class PhraseKerningCalibration extends HolidayAnimator {
 //        popStyle();
 //    }
 
-    public void addSprites(int t) {
+    public void addSprites(float t) {
         if (t == 0) {
             shapes.add(new Phrase(PHRASE1, PHRASE1_X_OFFSETS,
                     PHRASE1_Y_OFFSETS, (int) (0.8f * height)));
