+++ java/src/itp/comptypo/SnowflakeDemo.java
@@ -13,7 +13,7 @@ public class SnowflakeDemo extends HolidayAnimator {
         super(false, "Hy");
     }
 
-    public void addSprites(int t) {
+    public void addSprites(float t) {
         if (t == 0) {
             for (int i = 0; i < 20; i++) {
                 shapes.add(new Snowflake(random(width)));
